pyproject.toml
[project]
name = "comedy-store-transcribe"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pyqt6",
    "pyannote.audio>=4.0",
    "python-dotenv>=1.0",
    "torch>=2.3",
]

[dependency-groups]
dev = ["pytest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
